+++ src/TypeChecks/IsList.php
@@ -45,7 +45,7 @@ namespace GanbaroDigital\MissingBits\TypeChecks;
 
 use GanbaroDigital\MissingBits\Checks\Check;
 use GanbaroDigital\MissingBits\Checks\ListCheck;
-use GanbaroDigital\MissingBits\Checks\ListableCheck;
+use GanbaroDigital\MissingBits\Checks\ListCheckHelper;
 
 /**
  * do we have a valid PHP list?
@@ -53,7 +53,7 @@ use GanbaroDigital\MissingBits\Checks\ListableCheck;
 class IsList implements Check, ListCheck
 {
     // saves us having to implement inspectList() ourselves
-    use ListableCheck;
+    use ListCheckHelper;
 
     /**
      * can $list be safely (and sensibly) used in a foreach() loop?